organization-management-service/app/schemas/admin_schema.py
# app/schemas/admin_schema.py
from pydantic import BaseModel, EmailStr

class AdminLogin(BaseModel):
    email: EmailStr
    password: str

class TokenOut(BaseModel):
    access_token: str
    token_type: str = "bearer"
