organization-management-service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pymongo==4.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
